icnss: Don't create debugfs when !CONFIG_ICNSS_DEBUG
The debugfs related codes are guarded using CONFIG_ICNSS_DEBUG
macro. Below that, there is another chunk of codes for debugfs.
If we disable CONFIG_DEBUG_FS, it will still execute icnss_debugfs_create()
because of the else-condition of the ifdeffery. Therefore, remove the
irrelevant function and make ICNSS_DEBUG depend on DEBUG_FS.
Guard the debugfs bits accordingly.

Change-Id: I15e685774a84bc9d194fa688996804984dde1349
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>